seguranca: adiciona verificacao de tenant em taskcontroller (update/destroy)

+++ app/Controllers/TaskController.php
@@ -193,6 +193,11 @@ class TaskController extends Controller
             $data['due_date'] = str_replace('T', ' ', $this->inputRaw('due_date')) . ':00';
 
         $taskModel = new Task();
+        $task = $taskModel->findById($id);
+        if (!$task) {
+            $this->redirect('/tasks');
+            return;
+        }
         $taskModel->update($id, $data);
 
         // Se for requisição AJAX, retorna JSON; senão, redireciona
@@ -208,6 +213,11 @@ class TaskController extends Controller
     {
         $id = (int) ($params['id'] ?? 0);
         $taskModel = new Task();
+        $task = $taskModel->findById($id);
+        if (!$task) {
+            $this->redirect('/tasks');
+            return;
+        }
         $taskModel->delete($id);
 
         if (!empty($_SERVER['HTTP_X_REQUESTED_WITH'])) {
